+++ AdColony/src/main/java/com/mopub/mobileads/AdColonyAdapterConfiguration.java
@@ -10,7 +10,9 @@ import androidx.annotation.Nullable;
 
 import com.adcolony.sdk.AdColony;
 import com.adcolony.sdk.AdColonyAppOptions;
+import com.adcolony.sdk.AdColonyAdOptions;
 import com.mopub.common.BaseAdapterConfiguration;
+import com.mopub.common.DataKeys;
 import com.mopub.common.MoPub;
 import com.mopub.common.OnNetworkInitializationFinishedListener;
 import com.mopub.common.Preconditions;
@@ -31,7 +33,6 @@ public class AdColonyAdapterConfiguration extends BaseAdapterConfiguration {
     // Adapter's keys
     private static final String ADAPTER_NAME = AdColonyAdapterConfiguration.class.getSimpleName();
     private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
-    private static final String BIDDING_TOKEN = "1";
     private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
 
     // AdColony-specific keys (do not modify)
@@ -39,6 +40,7 @@ public class AdColonyAdapterConfiguration extends BaseAdapterConfiguration {
     protected static final String ALL_ZONE_IDS_KEY = "allZoneIds";
     protected static final String CLIENT_OPTIONS_KEY = "clientOptions";
     protected static final String ZONE_ID_KEY = "zoneId";
+    protected static final String KEY_ADCOLONY_BID_RESPONSE = "adm";
 
     // AdColony Default values
     protected static final String DEFAULT_ZONE_ID = "YOUR_CURRENT_ZONE_ID";
@@ -54,7 +56,7 @@ public class AdColonyAdapterConfiguration extends BaseAdapterConfiguration {
     @Nullable
     @Override
     public String getBiddingToken(@NonNull final Context context) {
-        return BIDDING_TOKEN;
+        return AdColony.collectSignals();
     }
 
     @NonNull
@@ -243,5 +245,52 @@ public class AdColonyAdapterConfiguration extends BaseAdapterConfiguration {
         Arrays.sort(newZones);
 
         return !Arrays.equals(previousZones, newZones);
+    }
+
+    protected AdColonyAdOptions getBannerAdOptionsFromExtras(Map<String, String> extras) {
+        return getAdOptionsFromExtras(extras, null);
+    }
+
+    protected AdColonyAdOptions getInterstitialAdOptionsFromExtras(Map<String, String> extras) {
+        return getAdOptionsFromExtras(extras, null);
+    }
+
+    protected AdColonyAdOptions getRewardAdOptionsFromExtras(Map<String, String> extras, String mAdUnitId) {
+        return getAdOptionsFromExtras(extras, mAdUnitId);
+    }
+
+    private AdColonyAdOptions getAdOptionsFromExtras(Map<String, String> extras, String mAdUnitId) {
+        final AdColonyAdOptions options = new AdColonyAdOptions();
+
+        // Rewarded Video only
+        if (mAdUnitId != null) {
+            options.enableConfirmationDialog(getConfirmationDialogFromSettings(mAdUnitId));
+            options.enableResultsDialog(getResultsDialogFromSettings(mAdUnitId));
+        }
+
+        // Forward ADM field contents to AdColony for Advanced Bidding requests.
+        // This is a requirement of AdColony Advanced Bidding 3.0 spec.
+        if (extras != null && extras.containsKey(DataKeys.ADM_KEY)
+                && !TextUtils.isEmpty(extras.get(DataKeys.ADM_KEY))) {
+            final String adMarkup = extras.get(DataKeys.ADM_KEY);
+
+            if (adMarkup != null) {
+                options.setOption(KEY_ADCOLONY_BID_RESPONSE, adMarkup);
+            }
+        }
+
+        return options;
+    }
+
+    private boolean getConfirmationDialogFromSettings(String mAdUnitId) {
+        final AdColonyRewardedVideo.AdColonyInstanceMediationSettings settings =
+                MoPubRewardedVideoManager.getInstanceMediationSettings(AdColonyRewardedVideo.AdColonyInstanceMediationSettings.class, mAdUnitId);
+        return settings != null && settings.isWithConfirmationDialog();
+    }
+
+    private boolean getResultsDialogFromSettings(String mAdUnitId) {
+        final AdColonyRewardedVideo.AdColonyInstanceMediationSettings settings =
+                MoPubRewardedVideoManager.getInstanceMediationSettings(AdColonyRewardedVideo.AdColonyInstanceMediationSettings.class, mAdUnitId);
+        return settings != null && settings.isWithResultsDialog();
     }
 }